+++ app/Http/Controllers/SubjectController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Subject;
 use Illuminate\Http\Request;
+use Auth;
 use Haruncpi\LaravelIdGenerator\IdGenerator;
 use Illuminate\Contracts\Validation\Validator;
 
@@ -14,11 +15,19 @@ class SubjectController extends Controller
      */
     public function index()
     {
+        if(Auth::user()->role == 'Teacher'){
+
         $subjects = Subject::latest()->get();
        
         return view('subjects',[
             'subjects' => $subjects
         ]);
+    }
+    else {
+        $data['title'] = '404';
+        $data['name'] = 'Page not found';
+        return response()->view('errors.404',$data,404);
+    }
     }
 
     /**
@@ -35,6 +44,7 @@ class SubjectController extends Controller
 
     public function store(Request $request)
     {
+        if(Auth::user()->role == 'Teacher'){
         $request->validate([
             'subject' => ['required', 'string', 'max:255', 'unique:subjects'],
         ]);
@@ -47,7 +57,12 @@ class SubjectController extends Controller
         $subject->subject = request('subject'); 
         $subject->save(); 
         return back()->with('subAddmessage', 'Subject added successfully');
-
+    }
+    else {
+        $data['title'] = '404';
+        $data['name'] = 'Page not found';
+        return response()->view('errors.404',$data,404);
+    }
     }
 
     /**
@@ -79,8 +94,15 @@ class SubjectController extends Controller
      */
     public function destroy(string $id)
     {
+    if(Auth::user()->role == 'Teacher'){
         $subject = Subject::find($id);
         $subject->delete();
         return redirect()->back()->with('deleted', 'Subject Deleted Successfully');
+    }
+    else {
+        $data['title'] = '404';
+        $data['name'] = 'Page not found';
+        return response()->view('errors.404',$data,404);
+    }
     }
 }
